% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-species.R
\name{decode}
\alias{decode}
\title{Decode}
\usage{
decode(speciesId, languageCode)
}
\value{
Data frame from all species & ids in defined language
}
\description{
Returns a list of names of the given id.
(Multiple names if there are multiple names for the tree species in the specified language.)
}
\examples{
decode("cDD", "en")
decode("cDD", "de")
decode("cDD", "es")
}
