% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-species.R
\name{getColorFromId}
\alias{getColorFromId}
\title{Color from species id}
\usage{
getColorFromId(speciesId)
}
\arguments{
\item{speciesId}{Id of the species}
}
\value{
character: hexadecimal color
}
\description{
Returns a color code for given species id
}
\examples{
getColorFromId("cDD")
}
