% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-species.R
\name{getRandomSpeciesId}
\alias{getRandomSpeciesId}
\title{Random species id}
\usage{
getRandomSpeciesId(languageCode)
}
\arguments{
\item{languageCode}{Language code ("de", "en", "es")}
}
\value{
3 character species id
}
\description{
Returns random generated species id
}
\examples{
getRandomSpeciesId("en")
getRandomSpeciesId("de")
getRandomSpeciesId("es")
}
