% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-species.R
\name{getSpeciesLength}
\alias{getSpeciesLength}
\title{Species Length}
\usage{
getSpeciesLength(languageCode)
}
\value{
integer
}
\description{
Number of tree species available in the specified language
}
\examples{
getColorFromId("cDD")
}
