% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-species.R
\name{encode}
\alias{encode}
\title{Encode}
\usage{
encode(name, languageCode)
}
\value{
3 character species id
}
\description{
Returns a 3 character species id from a given species name.
}
\examples{
encode("guatemalan fir", "en")
encode("Amerikanische Kastanie", "de")
encode("cedro oaxqueño", "es")
}
